Small changes to make sure everything is working properly.
During tSend to Elsa it seems that if(client.available()) is never true and can be removed. We somehow do get an answer from the server but client.available() just doesn't see it. I however kept this if as it prevents a crash on runtime that happens when Frida loses Wifi connection while expecting an answer from the server.